Updated to use latest version of o-link-list
o-link-list has removed it's font dependency, so we should use that here.
This commit also changes the demo dependency from o-typography to o-fonts.
We could have used o-typography, but it doesn't have to correct classes defined
for the styles we need, so we're actually only using the o-fonts functionality
as provided by proxy through o-typography.