Fix demo/use o--if-js
- use o--if-js and o--if-no-js instead of .core and .enhanced
- The way to specify the container height after which the the content should be
show/hideable is by setting a max-height. In our case we want to show/hide
everything so that max-height is 0. o-expander then ought to override that when
by adding o-expander__content--expanded which is styled to have max-height: none.
However this isn't working because of specificity. Adding styling for
`.o--if-js .o-section-head .o-expander__content--expanded` fixes this though it
does seem a bit noddy.